Refactor logging function
